+++ Space-Simulator/Stars.cpp
@@ -8,6 +8,6 @@
 
 #include "Stars.hpp"
 
-Star::Star(std::string name_, long id_, long double mass_, long double velocity_) {
+Star::Star(std::string name_, int id_, long double mass_, long double velocity_) {
     name = name_;
 }
